requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv==1.0.0
pydantic>=2.0.0
pyyaml>=6.0
colorama>=0.4.6
matplotlib>=3.5.0
tabulate>=0.9.0